pyproject.toml
[tool.poetry]
name = "Pink FastAPI Workshop Web App START"
version = "0.1.0"
description = ""
authors = ["Tilda Barth <tilda.barth@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.1"
Jinja2 = "^3.0.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
